pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rvgen"
version = "0.0.1"
description = "Pure-Python re-implementation of riscv-dv random instruction generator."
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[project.scripts]
rvgen = "rvgen.cli:main"

[tool.setuptools.packages.find]
include = ["rvgen*"]
exclude = ["tests*", "research*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
